refactor(tui): extract XtermTaskPty snapshot refresh into XtermSnapshotEngine
pty-xterm-base.ts was pinned at the 500-line file-size cap with zero
headroom. Move the scrollback cache, anchor marker, and snapshot rebuild
logic into a new pty-xterm-snapshot.ts module so the base class focuses on
transport and lifecycle.

No behavior change: all terminal snapshot/scrollback/cache tests pass.
pty-xterm-base.ts drops from 500 to 388 lines; the new engine is 149 lines.
